feat: Change the passage in the app, handle bold style and text color (#7)
Changed from Lorep Ipsum to some usage instructions.

At the same time, Tailwind prose font-weight and color gets overridden. The weight of bold style is set to normal weight + 300 (or max weight if adding 300 is not possible). Text color of the text is set to black or white depending on preferred color scheme.